Working backend and introduction of number rating
In this sprint, I edited a page adding a number rating which I intend to update to a star rating system in a future sprint. I also completed the backend which will insert whatever is put into the form into the database. This currently works if the user ID constraint is commented out. This is because the user ID will be identified by the user logged on the session when the other pages are connected. I have also added a place holder drop down to select the product ID. This will be removed in a later sprint.