Created a try block to query the database to get the email
I have created a try block to query the database to look in the user table for the email. I have stored the user id of the user that is logged in the session into a variable. That variable will be used as a reference in the query so that we get the right users email