Completely working Backend with comments
The backend of the upload products page works. It allows the user to upload their own products with information and price. Users can also add images of the product which gets stored on the server and the image address gets added to the database. Next step is to implement the admin approve feature to allow the admin to approve the product in order for it to be seen on the products page